Updated Javier to actually properly accept settings. Javier now actually accepts settings through default and server-specifics

+++ Internals/serverRelated.py
@@ -1,7 +1,10 @@
 """bunch of server related functions, from grabbing servercount to starting servers"""
 import os
 import subprocess
+import threading
 import zipfile
+from Internals import jdb
+jdb.deploy()
 def folders(dir = os.getcwd()):
     """gets all folders that have a .jar file in it from passed in directory.\n if nothing is passed, will check CWD"""
     servers = []
@@ -27,30 +30,38 @@ def folders(dir = os.getcwd()):
     return servers
 
 def runServer(server, dire, RAM):
-
-    dire = dire.replace("\\", "/")
-    file =[]
-    for item in os.listdir(f"{dire}/{server}"):
-        if item[-4:] == '.jar':
-            file.append(item)
-    if len(file) > 1:
-        for item in file:
-            check = zipfile.ZipFile(f"{dire}/{server}/{item}",'r')
-            check = check.open('META-INF/MANIFEST.MF','r')
-            check = check.readlines()
-            for line in check:
-                if 'net.minecraft.server.MinecraftServer' in str(line):
-                        file.remove(item)
-    jar = f"{dire}/{server}/{file[0]}"
+    if jdb.readServerValue(server, "JARName") == None:
+        dire = dire.replace("\\", "/")
+        file =[]
+        for item in os.listdir(f"{dire}/{server}"):
+            if item[-4:] == '.jar':
+                file.append(item)
+        if len(file) > 1:
+            for item in file:
+                check = zipfile.ZipFile(f"{dire}/{server}/{item}",'r')
+                check = check.open('META-INF/MANIFEST.MF','r')
+                check = check.readlines()
+                for line in check:
+                    if 'net.minecraft.server.MinecraftServer' in str(line):
+                            file.remove(item)
+        jar = f"{dire}/{server}/{file[0]}"
+        jdb.updateServerValue(server, "JARName", jar)
+    else:
+        jar = jdb.readServerValue(server, "JARName")
     universe = f"{dire}/{server}"
-    java = "java"
+    if jdb.readServerValue(server, "JavaFilePath") != '':
+        java = jdb.readServerValue(server,"JavaFilePath")
+    elif jdb.readSettingValue("DefaultJava") != None:
+        java = jdb.readSettingValue("DefaultJava")
+    else:
+        java = "java"
     operating = os.name
     print(operating)
     if operating == "nt":
         cmd = (f'start cmd /C {java} -Xmx{RAM}G -Xms256M -jar "{jar}" nogui')
         subprocess.run((cmd), shell=True, cwd=universe)
-        #cmd = (f"{java}", f"-Xmx{RAM}G", "-Xms256M", "-jar", jar, "nogui")  # why doesn't it work!!
-        #subprocess.Popen((cmd), shell=True, cwd=universe, creationflags=subprocess.CREATE_NEW_CONSOLE
+            #cmd = (f"{java}", f"-Xmx{RAM}G", "-Xms256M", "-jar", jar, "nogui")  # why doesn't it work!!
+            #subprocess.Popen((cmd), shell=True, cwd=universe, creationflags=subprocess.CREATE_NEW_CONSOLE
     else:      #xterm -e    # MacOS hates this.  will have to determine a workaround for Mac, eventually.
-        cmd = (f"xterm -e '{java}' -Xmx{RAM}G -Xms256M -jar '{jar}' nogui") # probably will have it limited to 1 server on mac :L
+        cmd = (f"xterm -e '{java}' -Xmx{RAM}G -Xms256M -jar '{jar}' nogui")
         subprocess.run((cmd), shell=True, cwd=universe)
